requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-ninja==0.17.0
pydantic==1.9.0
sqlparse==0.4.2
typing-extensions==4.0.1
